+++ js/eme_admin_sendmails.js
@@ -513,6 +513,7 @@ jQuery(document).ready( function($) {
             selecting: true, // Enable selecting
             multiselect: true, // Allow multiple selecting
             selectingCheckboxes: true, // Show checkboxes on first column
+		multiSorting: true,
             defaultSorting: 'sent_datetime DESC',
             actions: {
                 listAction: ajaxurl,
@@ -662,6 +663,7 @@ jQuery(document).ready( function($) {
             selecting: true, // Enable selecting
             multiselect: true, // Allow multiple selecting
             selectingCheckboxes: true, // Show checkboxes on first column
+		multiSorting: true,
             defaultSorting: 'planned_on DESC, name',
             actions: {
                 listAction: ajaxurl,
@@ -776,6 +778,7 @@ jQuery(document).ready( function($) {
             selecting: true, // Enable selecting
             multiselect: true, // Allow multiple selecting
             selectingCheckboxes: true, // Show checkboxes on first column
+		multiSorting: true,
             defaultSorting: 'planned_on DESC, name',
             actions: {
                 listAction: ajaxurl,